requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
googleapis-common-protos==1.60.0
grpclib==0.4.5
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
multidict==6.0.4
Pillow==10.0.0
protobuf==4.24.1
requests==2.31.0
typing_extensions==4.7.1
urllib3==2.0.4
viam-sdk==0.4.8
